% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_eventbrite_function.R
\name{call_eventbrite}
\alias{call_eventbrite}
\title{Call Eventbrite}
\usage{
call_eventbrite(event_id = NA, query = "events", sub_query = NA,
  second_id = NA, token = NA)
}
\arguments{
\item{event_id}{the unique id that every Eventbrite event has}

\item{query}{the top-level query. Defaults to "events". Other query entries could be "users" or "orders".}

\item{sub_query}{the particular information wanted from the query. If query = "events", then sub_query could be "attendees", "users", "orders", "display_settings", "ticket_classes", "canned_questions", "questions", "discounts", "public_discounts", "access_codes", "transfers", "teams", "ticket_groups", "ticket_classes", "ticket_buyer_settings"}

\item{token}{the API token needed to call the API. Go here to create your own app and get an API token: http://www.eventbrite.com/myaccount/apps/}
}
\description{
This function compiles the url required to call the Eventbrite API based on a few parameters.
call_eventbrite returns the URL as a string, which should be used with get_eb_event()
function relies on "httr" package and will install if it doesn't exist
}
\examples{
url = call_eventbrite(event_id = event_id, sub_query = "orders", token = token) # will return the url for a given event's orders/registrations (using event_id)
event = get_eb_event(url)
orders = get_eb_orders(event)
}
