% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_query_function.R
\name{eb_query}
\alias{eb_query}
\title{Get Eventbrite object}
\usage{
eb_query(url)
}
\arguments{
\item{url}{the Eventbrite API url to call. This URL can be generated with call_eventbrite}
}
\description{
This function compiles extracts the Eventbrite's event data from its original JSON form.
function relies on "jsonlite" package and will install if it doesn't exist
}
\examples{
url = call_eventbrite(event_id = event_id, sub_query = "orders", token = token) # will return the url for a given event's orders/registrations (using event_id)
event = eb_query(url)
orders = get_eb_orders(event)
}
