% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eb_attendees_function.R
\name{get_eb_attendees}
\alias{get_eb_attendees}
\title{Get Eventbrite attendees}
\usage{
get_eb_attendees(event, names.only = TRUE)
}
\arguments{
\item{event}{the event object returned from get_eb_attendees(url)}

\item{names.only}{Default set TRUE. To return a clean data.frame of attendees, which only contains names of registrations. If user wants all data in output, they can set names.only=FALSE}
}
\description{
This function returns the Eventbrite attendees of an event. The function only works on Events that hold attendees data. As such, it is designed to be used on the output of eb_query() (see example).
It will return a dataframe of attendees of the course including a column giving a status on if they were "checked-in"
}
\examples{
attendees = eb_query(event_id = event_id, sub_query = "attendees") \%>\% get_eb_orders(attendees)
}
